# be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/day-states": {
      "get": {
        "operationId": "DayStatesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayStateResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all day states for the current user",
        "tags": [
          "Day States"
        ]
      },
      "post": {
        "operationId": "DayStatesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDayStateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayStateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new day state",
        "tags": [
          "Day States"
        ]
      }
    },
    "/api/v1/day-states/from-recommendation": {
      "post": {
        "operationId": "DayStatesController_createFromRecommendation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDayStateFromRecommendationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayStateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Unknown recommendation key"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a day state from a recommendation",
        "tags": [
          "Day States"
        ]
      }
    },
    "/api/v1/day-states/{id}": {
      "patch": {
        "operationId": "DayStatesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDayStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayStateResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Day state not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a day state",
        "tags": [
          "Day States"
        ]
      },
      "delete": {
        "operationId": "DayStatesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Day state deleted"
          },
          "404": {
            "description": "Day state not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a day state",
        "tags": [
          "Day States"
        ]
      }
    },
...
    "/api/v1/days/{date}": {
      "put": {
        "operationId": "DaysController_upsert",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "path",
            "description": "Date in YYYY-MM-DD format",
            "schema": {
              "example": "2024-01-15",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertDayDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Day state not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set or update the day state for a specific date",
        "tags": [
          "Days"
        ]
      }
    },
    "/api/v1/days/{date}/location": {
      "patch": {
        "operationId": "DaysController_updateLocation",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "path",
            "description": "Date in YYYY-MM-DD format",
            "schema": {
              "example": "2024-01-15",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDayLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Future date or invalid coordinates"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update or clear the location for a specific date",
        "tags": [
          "Days"
        ]
      }
    },
    "/api/v1/days": {
      "get": {
        "operationId": "DaysController_findAll",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "description": "Start of date range (YYYY-MM-DD)",
            "schema": {
              "example": "2024-01-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "description": "End of date range (YYYY-MM-DD)",
            "schema": {
              "example": "2024-01-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get days with their states for a date range",
        "tags": [
          "Days"
        ]
      }
    },
...
  },
  "info": {
    "title": "LifeSpan API",
    "description": "Backend API for Life Timeline — build a visual timeline of your life",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DayStateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "isSystem": {
            "type": "boolean"
          },
          "order": {
            "type": "number"
          },
          "score": {
            "type": "number",
            "description": "Mood intensity score (0-10)"
          }
        },
        "required": [
          "id",
          "name",
          "color",
          "isSystem",
          "order",
          "score"
        ]
      },
      "CreateDayStateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Great"
          },
          "color": {
            "type": "string",
            "example": "#4CAF50",
            "description": "Hex color code"
          },
          "order": {
            "type": "number",
            "example": 0,
            "description": "Display order"
          },
          "score": {
            "type": "number",
            "example": 7,
            "description": "Mood intensity score (0-10)"
          }
        },
        "required": [
          "name",
          "color",
          "score"
        ]
      },
      "CreateDayStateFromRecommendationDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "great",
            "description": "Recommendation key"
          },
          "name": {
            "type": "string",
            "example": "Great",
            "description": "Translated display name"
          }
        },
        "required": [
          "key",
          "name"
        ]
      },
      "UpdateDayStateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Great"
          },
          "color": {
            "type": "string",
            "example": "#4CAF50"
          },
          "order": {
            "type": "number",
            "example": 0
          },
          "score": {
            "type": "number",
            "example": 7,
            "description": "Mood intensity score (0-10)"
          }
        }
      },
...
      "UpsertDayDto": {
        "type": "object",
        "properties": {
          "dayStateId": {
            "type": "string",
            "example": "day-state-id",
            "nullable": true,
            "description": "Day state ID. Pass null to clear the day state."
          },
          "mainMediaId": {
            "type": "string",
            "example": "media-uuid",
            "nullable": true,
            "description": "ID of the media item to use as the day cover. Pass null to clear."
          },
          "comment": {
            "type": "string",
            "example": "Had a great day!",
            "nullable": true,
            "description": "Day comment (max 500 chars). Pass null to clear."
          }
        }
      },
      "DayStateSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "color"
        ]
      },
      "DayMediaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "s3Key": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "description": "Presigned URL for reading the media file (null if generation failed)",
            "nullable": true
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "s3Key",
          "url",
          "fileName",
          "contentType",
          "size",
          "createdAt"
        ]
      },
      "DayResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "example": "2024-01-15"
          },
          "dayState": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/DayStateSummaryDto"
              }
            ]
          },
          "mainMediaId": {
            "type": "string",
            "nullable": true,
            "example": "media-uuid",
            "description": "ID of the cover media"
          },
          "locationName": {
            "type": "string",
            "nullable": true,
            "example": "Kyiv, Ukraine",
            "description": "Location name"
          },
          "latitude": {
            "type": "number",
            "nullable": true,
            "example": 50.4501,
            "description": "Latitude"
          },
          "longitude": {
            "type": "number",
            "nullable": true,
            "example": 30.5234,
            "description": "Longitude"
          },
          "comment": {
            "type": "string",
            "nullable": true,
            "example": "Had a great day!",
            "description": "Day comment"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayMediaResponseDto"
            }
          }
        },
        "required": [
          "id",
          "date",
          "dayState",
          "mainMediaId",
          "locationName",
          "latitude",
          "longitude",
          "comment",
          "media"
        ]
      },
      "UpdateDayLocationDto": {
        "type": "object",
        "properties": {
          "locationName": {
            "type": "string",
            "example": "Kyiv, Ukraine",
            "nullable": true,
            "description": "Location name (max 120 chars). Pass null to clear."
          },
          "latitude": {
            "type": "number",
            "example": 50.4501,
            "nullable": true,
            "description": "Latitude (-90 to 90). Pass null to clear."
          },
          "longitude": {
            "type": "number",
            "example": 30.5234,
            "nullable": true,
            "description": "Longitude (-180 to 180). Pass null to clear."
          }
        }
      },
...
    }
  }
}